requirements.txt
alpaca-py==0.43.2
altair==6.0.0
annotated-types==0.7.0
anywidget==0.10.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.15.0
cycler==0.12.1
dateparser==1.4.0
decorator==5.2.1
dill==0.4.1
executing==2.2.1
fonttools==4.62.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.3
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
msgpack==1.1.2
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.19.0
numba==0.65.0
numpy==2.4.4
packaging==26.0
pandas==2.3.3
parso==0.8.6
peewee==4.0.4
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psycopg2-binary==2.9.11
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.49
sseclient-py==1.9.0
stack-data==0.6.3
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vectorbt==0.28.5
wcwidth==0.6.0
websockets==16.0
widgetsnbextension==4.0.15
yfinance==1.2.1
